third_party/blink/web_tests/platform/generic/media/W3C/audio/readyState/readyState_initial-expected.txt

audio.readyState - default state

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS "0" is "0"
PASS successfullyParsed is true

TEST COMPLETE


spec reference
